PYIC-7797: move ticf cri service into library.
