+++ libs/ticf-cri-service/src/main/java/uk/gov/di/ipv/core/library/ticfcriservice/dto/TicfCriDto.java
@@ -1,4 +1,4 @@
-package uk.gov.di.ipv.core.processcandidateidentity.dto;
+package uk.gov.di.ipv.core.library.ticfcriservice.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import uk.gov.di.ipv.core.library.enums.Vot;
