+++ libs/ticf-cri-service/src/main/java/uk/gov/di/ipv/core/library/ticfcriservice/exception/TicfCriHttpResponseException.java
@@ -1,4 +1,4 @@
-package uk.gov.di.ipv.core.processcandidateidentity.exception;
+package uk.gov.di.ipv.core.library.ticfcriservice.exception;
 
 public class TicfCriHttpResponseException extends Exception {
     public TicfCriHttpResponseException(String message) {
